+++ packages/frontend/src/components/legislator/index.tsx
@@ -1,11 +1,8 @@
 'use client'
 import React, { useState, useEffect } from 'react'
-import { useRouter } from 'next/navigation'
+import { useRouter, useSearchParams } from 'next/navigation'
 // @twreporter
-import {
-  DesktopAndAbove,
-  TabletAndBelow,
-} from '@twreporter/react-components/lib/rwd'
+import { TabletAndBelow } from '@twreporter/react-components/lib/rwd'
 // components
 import LegislatorInfo from '@/components/legislator/legislator-info'
 import LegislatorStatistics from '@/components/legislator/legislator-statistics'
@@ -46,8 +43,19 @@ const Legislator: React.FC<LegislatorProps> = ({
   currentMeetingSession,
 }) => {
   const router = useRouter()
+  const searchParams = useSearchParams()
   const [filterCount, setFilterCount] = useState<number>(0)
   const [isLegislatorActive, setIsLegislatorActive] = useState<boolean>(false)
+  const [isLoading, setIsLoading] = useState<boolean>(false)
+
+  useEffect(() => {
+    setIsLoading(false)
+  }, [searchParams])
+
+  const { legislator, topics, speechesByTopic } = useLegislatorData(
+    legislatorData,
+    topicsData
+  )
 
   const {
     isFilterOpen,
@@ -57,12 +65,11 @@ const Legislator: React.FC<LegislatorProps> = ({
     legislativeMeetingState,
     legislativeMeetingSessionState,
     filterOptions,
-  } = useLegislativeMeetingFilters(currentMeetingTerm, currentMeetingSession)
-
-  const { legislator, topics, speechesByTopic } = useLegislatorData(
-    legislatorData,
-    topicsData
-  )
+  } = useLegislativeMeetingFilters({
+    legislatorSlug: legislator.slug,
+    currentMeetingTerm,
+    currentMeetingSession,
+  })
 
   const openFilter = () => {
     setIsFilterOpen((prev) => !prev)
@@ -97,6 +104,7 @@ const Legislator: React.FC<LegislatorProps> = ({
         sessionTermValue
       )}`
       router.push(newUrl)
+      setIsLoading(true)
       setIsFilterOpen(false)
     } catch (error) {
       console.error('Error processing filter values:', error)
@@ -122,11 +130,7 @@ const Legislator: React.FC<LegislatorProps> = ({
         setFilterCount(currentMeetingSession.length)
       }
     }
-  }, [
-    currentMeetingTerm,
-    currentMeetingSession,
-    legislativeMeetingSessionState,
-  ])
+  }, [currentMeetingTerm, currentMeetingSession])
 
   useEffect(() => {
     // if the legislator's meeting term matches the latest meeting term, set isLegislatorActive to true
@@ -148,35 +152,34 @@ const Legislator: React.FC<LegislatorProps> = ({
         filterCount={filterCount}
         onFilterClick={openFilter}
       >
-        <DesktopAndAbove>
-          <DesktopContainer>
-            <DesktopAsideLeft>
-              <LegislatorInfo
-                legislator={legislator}
-                isLegislatorActive={isLegislatorActive}
+        <DesktopContainer>
+          <DesktopAsideLeft>
+            <LegislatorInfo
+              legislator={legislator}
+              isLegislatorActive={isLegislatorActive}
+            />
+            <Feedback />
+          </DesktopAsideLeft>
+          <DesktopAsideRight>
+            <LegislatorStatistics
+              committees={legislator.committees}
+              proposalSuccessCount={legislator.proposalSuccessCount}
+              meetingTermCount={legislator.meetingTermCount}
+              meetingTermCountInfo={legislator.meetingTermCountInfo}
+            />
+            <ListContainer>
+              <LegislatorList
+                isLoading={isLoading}
+                legislatorSlug={legislator.slug}
+                legislatorName={legislator.name}
+                topics={topics}
+                speechesByTopic={speechesByTopic}
+                currentMeetingTerm={currentMeetingTerm}
+                currentMeetingSession={currentMeetingSession}
               />
-              <Feedback />
-            </DesktopAsideLeft>
-            <DesktopAsideRight>
-              <LegislatorStatistics
-                committees={legislator.committees}
-                proposalSuccessCount={legislator.proposalSuccessCount}
-                meetingTermCount={legislator.meetingTermCount}
-                meetingTermCountInfo={legislator.meetingTermCountInfo}
-              />
-              <ListContainer>
-                <LegislatorList
-                  legislatorSlug={legislator.slug}
-                  legislatorName={legislator.name}
-                  topics={topics}
-                  speechesByTopic={speechesByTopic}
-                  currentMeetingTerm={currentMeetingTerm}
-                  currentMeetingSession={currentMeetingSession}
-                />
-              </ListContainer>
-            </DesktopAsideRight>
-          </DesktopContainer>
-        </DesktopAndAbove>
+            </ListContainer>
+          </DesktopAsideRight>
+        </DesktopContainer>
         <TabletAndBelow>
           <ContentBlock>
             <LegislatorInfo
@@ -191,6 +194,7 @@ const Legislator: React.FC<LegislatorProps> = ({
             />
             <ListContainer>
               <LegislatorList
+                isLoading={isLoading}
                 legislatorSlug={legislator.slug}
                 legislatorName={legislator.name}
                 topics={topics}